WIP Extract discovery from SpecClassDescriptor
